Made BankTableModel generic
